Move: CSLTS --> CJIBG35
moved from clts-project to cjibg35-project as this is the desired project to run the simulation.
To move this i cloned the workflow / modelsetup clean at the desired location. This is / was a good prove if everything works 'out-ofthe-box'.
At the same time I compiled as many needed software for the new OS at JUWELS as Im able to. (DWDlibgrib, Grib_API, INT2LM). TSMP has to be adjusted by AGI, which is in progress.

-) export_paths.ksh
-) clean up a bit