update generation of HISTORY.txt
-) since destructive git operations are available (e.g. git rebase
which leads to destroying or re-creating of the commit sha1) the git
commit sha1 is not sufficient to uniquely identify a commit. However,
even if the commit sha1 is recalculated information about the
authored date, the author and commit-message do remain. So I added
the last entire log entry to HISTORY.txt instead of commit sha1 only.
(If above information are not enough to track the correct commit,
you can forget everything...)
-) further add autodetecting of repo remote URL to make this mworkflow
more generic