add checkGitStatus()
-) This function does make sure that the working tree is clean
and each git commit used to run a simulaiton has its own tag.
The clean working tree is needed, as the correct setup is not
trackable if some files (or latest versions) are not tracked by git.
The seperat tag is needed as the commit-hash is changable and
not a 'perfect' indicator of the exact used git commit refering
to the correct simulation setup.