Generalize simulation length
The naming of sub-dirs within simres/ postpro/ monitoring/ etc.
is fix to YYYY_MM, as this workflow was developed from a climate
run, subdivided into several monthly simulations. However, this
is not suitable for other simulations types, like case-studies or
forecasts, which might be subdivided into daily or hourly
simulations. So a flexible naming scheme should be implemented,
using a dateString controled within ctrl/starter.sh.

- add using dateString to control naming of result sub-dirs
- use UTC for ALL date operations (date -u)
- use date norm ISO-8601
- remove hard-coded flaggs in namelists as e.g. __year__ and __moth__
- add a workaround for the post-processing (scripts within ctrl/postpro
are not using the new individual dateString, but everything is copyed
to the correct place automatically)